Enable building as a web extension
For this extension to work on VS Code Web (powering vscode.dev and
github.dev), it needs to be built as a Web extension. For this purpose,
I added a Webpack config and added entries to JSON files. I didn't need
to change the codebase because it doesn't depend to Node or to a
filesystem. I tested the changes at both local machine and vscode.dev.
The features work seamlessly.

I didn't touch the desktop extension build, for not breaking anyting.
It is compiled and packaged the same way as before.
In the future, making Nightlight only a Web extension might be useful
(Web extensions run on VS Code Desktop), removing the need for
maintaining two separate builds.